// src/types/demo.d.ts
export interface CustomEvent {
    name?: string
    dimmed?: boolean
    status: string
    date?: string
    color?: string
    icon?: string
    image?: string
}